% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{faces}
\alias{faces}
\title{Face data base from the matlab pnmf package}
\format{Faces are 32x32 images, dataset is 2049 rows,
1024 (=32x32) columns}
\usage{
faces
}
\description{
Originally from the FERET database, https://www.itl.nist.gov/iad/humanid/feret/.
}
\keyword{datasets}
