% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{NMFimage2df}
\alias{NMFimage2df}
\title{ggplot2 dataframe from NMF image data}
\usage{
NMFimage2df(nmfimage, rs, cs)
}
\arguments{
\item{nmfimage}{- data for NMF - each row is an image}

\item{rs}{- number of rows in each image}

\item{cs}{- number of cols in each image}
}
\value{
data frame for use with ggplot2
}
\description{
ggplot2 dataframe from NMF image data
}
\details{
transform a typical basis data frame to something that can be
displayed as images with ggplot2. Each image is a row and the dimensions
need to be passed in.
}
\examples{
data(faces)
j<-NMFimage2df(faces, 32, 32)
ggplot(filter(j, ID <= 16), aes(x=Var2, y=33-Var1, fill=Brightness)) +
geom_tile() + facet_wrap(~ID) + scale_fill_gradient(low="black", high="white")
}
