% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard.R
\name{PNMFARD}
\alias{PNMFARD}
\title{Automatic rank determination PNMF based on euclidean distance.}
\usage{
PNMFARD(
  X,
  nmfMod,
  tol = 1e-05,
  maxIter = 5000,
  verboseN = FALSE,
  zerotol = 1e-10
)
}
\arguments{
\item{X}{Input data matrix}

\item{nmfMod}{NMF model from the NMF package}

\item{tol}{tolerance for stopping criteria}

\item{maxIter}{Maximum number of iterations}

\item{verbose}{Print status messages}
}
\value{
Fitted NMF model, as defined in NMF package.
The "extra" slot contains sigma and wnorm
}
\description{
Automatic rank determination PNMF based on euclidean distance.
}
\details{
Implementation of "Automatic Rank Determination in Projective Nonnegative Matrix Factorization."
Zhirong Yang, Zhanxing Zhu, Erkki Oja. In the 9th International Conference on Latent Variable Analysis
and Signal Separation (LVA 2010), pages 514-521, St. Malo, France, 2010
Derived from
matlab code by Z. Yang, https://sites.google.com/site/zhirongyangcs/ardpnmf

sigma and wnorm are stored in the "extra" slot of the nmf object
}
\examples{
library(NMF)
setNMFMethod("PNMFARD", pNMF::PNMFARD)
mkD <- function(NOISE=TRUE) {
  n <- 1000 # rows
  counts <- c(30, 10, 20, 10, 15, 15) # samples
  syntheticNMF(n=n, r=counts, offset = NULL, noise = NOISE,
               factors = FALSE, seed = 99)
}
k<-mkD()
estim.r2 <- nmf(k, 16, method="PNMFARD", nrun=1, seed="nndsvd")
#wnorm and sigma in the extra slot
estim.r2@extra
}
\seealso{
PNMFARD
}
