% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{rss.rankplot}
\alias{rss.rankplot}
\title{Plotting for Frigyesi rank selection method}
\usage{
rss.rankplot(orig, random)
}
\arguments{
\item{orig}{result of a rank selection run on original data}

\item{random}{result of a rank selection run on randomised data}
}
\value{
a ggplot object
}
\description{
Plotting for Frigyesi rank selection method
}
\examples{
\dontrun{
  mkD <- function(NOISE=TRUE) {
  n <- 1000 # features
  counts <- c(30, 10, 20, 10, 15, 15) # samples
  syntheticNMF(n=n, r=counts, offset = NULL, noise = NOISE,
               factors = FALSE, seed = 99)
 }
 k<-mkD()
 V.random <- randomize(k)
 aheatmap(V.random)
 estim.r2 <- nmf(k, 2:20, nrun=30)
 estim.r2.random <- nmf(V.random, 2:20, nrun=30)
 rss.rankplot(estim.r2, estim.r2.random)
}
}
