% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnmf.R
\name{PNMF}
\alias{PNMF}
\title{Projective nonnegative matrix factorization}
\usage{
PNMF(X, nmfMod, tol = 1e-05, maxIter = 500, timeLimit = 300,
  checkDivergence = TRUE)
}
\arguments{
\item{checkDivergence}{}
}
\value{
Fitted NMF model, as defined in NMF package.
}
\description{
Projective nonnegative matrix factorization
}
\examples{
NMF::register(PNMF)

}
